ProfinetClient | Filter

+++ src/tests/unit/services/profinet_client/filter_threshold_float_test.rs
@@ -70,7 +70,7 @@ mod tests {
         let mut prev = 0.0;
         for (value, target) in testData {
             filter.add(value);
-            let diff = (prev as f32 - (value as f32)).abs();
+            let diff = (prev as f64 - (value as f64)).abs();
             if diff > threasold {
                 prev = value;
             }
@@ -140,7 +140,7 @@ mod tests {
         let mut prev = 0.0;
         for (value, target) in testData {
             filter.add(value);
-            let diff = (prev as f32 - (value as f32)).abs();
+            let diff = (prev as f64 - (value as f64)).abs();
             if diff > threasold {
                 prev = value;
             }
@@ -191,7 +191,7 @@ mod tests {
         let mut prev = 0.0;
         for (value, target) in testData {
             filter.add(value);
-            let diff = (prev as f32 - (value as f32)).abs();
+            let diff = (prev as f64 - (value as f64)).abs();
             if diff > threasold {
                 prev = value;
             }
@@ -261,7 +261,7 @@ mod tests {
         let mut prev = 0.0;
         for (value, target) in testData {
             filter.add(value);
-            let diff = (prev as f32 - (value as f32)).abs();
+            let diff = (prev as f64 - (value as f64)).abs();
             if diff > threasold {
                 prev = value;
             }
